+++ src/main/java/br/com/eric/telegram/kerobot/daos/ChatRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import br.com.eric.telegram.kerobot.models.ChatModel;
 
-public interface ChatRepository extends CrudRepository<ChatModel, Integer> {
+public interface ChatRepository extends CrudRepository<ChatModel, Long> {
 
 }
