+++ src/main/java/br/com/eric/telegram/kerobot/daos/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import br.com.eric.telegram.kerobot.models.UserModel;
 
-public interface UserRepository extends CrudRepository<UserModel, Integer> {
+public interface UserRepository extends CrudRepository<UserModel, Long> {
 
 	Optional<UserModel> findOneByUsername(String replaceAll);
